#269 [fix] viewModel에서 dispatcher 분리
