+++ app/src/main/java/hous/release/android/util/event/Reducer.kt
@@ -1,4 +1,4 @@
-package hous.release.android.util
+package hous.release.android.util.event
 
 fun interface Reducer<S, E> {
     fun dispatch(state: S, event: E): S
